[req-changes] Moved zerotier tasks and changed 'exc' to 'e' for consistency
